Update test and example dependencies

+++ Unicolour.Tests/DatasetEbnerFairchildTests.cs
@@ -47,14 +47,14 @@ public class DatasetEbnerFairchildTests
     ];
     
     [TestCaseSource(nameof(ReferenceHues))]
-    public void ReferenceHue(Unicolour colour, int expectedHue)
+    public void ReferenceHue(Unicolour colour, double expectedHue)
     {
         const double tolerance = 0.05;
         Assert.That(colour.Lchab.H, Is.EqualTo(expectedHue).Within(tolerance).Or.EqualTo(expectedHue + 360).Within(tolerance));
     }
 
     [TestCaseSource(nameof(GroupedByHue))]
-    public void GroupedHue(List<Unicolour> colours, int expectedHue, int expectedCount)
+    public void GroupedHue(List<Unicolour> colours, double expectedHue, int expectedCount)
     {
         Assert.That(colours.Count, Is.EqualTo(expectedCount));
